structured logger formatter prototype
