+++ src/libraries/Microsoft.Extensions.Logging.Console/src/ConsoleLogger.cs
@@ -3,34 +3,19 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Diagnostics;
-using System.Text;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.Extensions.Logging.Console
 {
     internal class ConsoleLogger : ILogger
     {
-        private static readonly string _loglevelPadding = ": ";
-        private static readonly string _messagePadding;
-        private static readonly string _newLineWithMessagePadding;
-
-        // ConsoleColor does not have a value to specify the 'Default' color
-        private readonly ConsoleColor? DefaultConsoleColor = null;
+        private readonly IEnumerable<ILogFormatter> _formatters;
 
         private readonly string _name;
         private readonly ConsoleLoggerProcessor _queueProcessor;
 
-        [ThreadStatic]
-        private static StringBuilder _logBuilder;
-
-        static ConsoleLogger()
-        {
-            var logLevelString = GetLogLevelString(LogLevel.Information);
-            _messagePadding = new string(' ', logLevelString.Length + _loglevelPadding.Length);
-            _newLineWithMessagePadding = Environment.NewLine + _messagePadding;
-        }
-
-        internal ConsoleLogger(string name, ConsoleLoggerProcessor loggerProcessor)
+        internal ConsoleLogger(string name, ConsoleLoggerProcessor loggerProcessor, IEnumerable<ILogFormatter> formatters)
         {
             if (name == null)
             {
@@ -39,6 +24,7 @@ namespace Microsoft.Extensions.Logging.Console
 
             _name = name;
             _queueProcessor = loggerProcessor;
+            _formatters = formatters;
         }
 
         internal IExternalScopeProvider ScopeProvider { get; set; }
@@ -67,161 +53,11 @@ namespace Microsoft.Extensions.Logging.Console
 
         public virtual void WriteMessage(LogLevel logLevel, string logName, int eventId, string message, Exception exception)
         {
-            var format = Options.Format;
-            Debug.Assert(format >= ConsoleLoggerFormat.Default && format <= ConsoleLoggerFormat.Systemd);
-
-            var logBuilder = _logBuilder;
-            _logBuilder = null;
-
-            if (logBuilder == null)
-            {
-                logBuilder = new StringBuilder();
-            }
-
-            LogMessageEntry entry;
-            if (format == ConsoleLoggerFormat.Default)
-            {
-                entry = CreateDefaultLogMessage(logBuilder, logLevel, logName, eventId, message, exception);
-            }
-            else if (format == ConsoleLoggerFormat.Systemd)
-            {
-                entry = CreateSystemdLogMessage(logBuilder, logLevel, logName, eventId, message, exception);
-            }
-            else
-            {
-                entry = default;
-            }
+            var formatter = _formatters.Single(f => f.Name == (Options.Formatter ?? Options.Format.ToString()));
+            var entry = formatter.Format(logLevel, logName, eventId, message, exception);
             _queueProcessor.EnqueueMessage(entry);
-
-            logBuilder.Clear();
-            if (logBuilder.Capacity > 1024)
-            {
-                logBuilder.Capacity = 1024;
-            }
-            _logBuilder = logBuilder;
         }
 
-        private LogMessageEntry CreateDefaultLogMessage(StringBuilder logBuilder, LogLevel logLevel, string logName, int eventId, string message, Exception exception)
-        {
-            // Example:
-            // INFO: ConsoleApp.Program[10]
-            //       Request received
-
-            var logLevelColors = GetLogLevelConsoleColors(logLevel);
-            var logLevelString = GetLogLevelString(logLevel);
-            // category and event id
-            logBuilder.Append(_loglevelPadding);
-            logBuilder.Append(logName);
-            logBuilder.Append('[');
-            logBuilder.Append(eventId);
-            logBuilder.AppendLine("]");
-
-            // scope information
-            GetScopeInformation(logBuilder, multiLine: true);
-
-            if (!string.IsNullOrEmpty(message))
-            {
-                // message
-                logBuilder.Append(_messagePadding);
-
-                var len = logBuilder.Length;
-                logBuilder.AppendLine(message);
-                logBuilder.Replace(Environment.NewLine, _newLineWithMessagePadding, len, message.Length);
-            }
-
-            // Example:
-            // System.InvalidOperationException
-            //    at Namespace.Class.Function() in File:line X
-            if (exception != null)
-            {
-                // exception message
-                logBuilder.AppendLine(exception.ToString());
-            }
-
-            string timestamp = null;
-            var timestampFormat = Options.TimestampFormat;
-            if (timestampFormat != null)
-            {
-                var dateTime = GetCurrentDateTime();
-                timestamp = dateTime.ToString(timestampFormat);
-            }
-
-            return new LogMessageEntry(
-                message: logBuilder.ToString(),
-                timeStamp: timestamp,
-                levelString: logLevelString,
-                levelBackground: logLevelColors.Background,
-                levelForeground: logLevelColors.Foreground,
-                messageColor: DefaultConsoleColor,
-                logAsError: logLevel >= Options.LogToStandardErrorThreshold
-            );
-        }
-
-        private LogMessageEntry CreateSystemdLogMessage(StringBuilder logBuilder, LogLevel logLevel, string logName, int eventId, string message, Exception exception)
-        {
-            // systemd reads messages from standard out line-by-line in a '<pri>message' format.
-            // newline characters are treated as message delimiters, so we must replace them.
-            // Messages longer than the journal LineMax setting (default: 48KB) are cropped.
-            // Example:
-            // <6>ConsoleApp.Program[10] Request received
-
-            // loglevel
-            var logLevelString = GetSyslogSeverityString(logLevel);
-            logBuilder.Append(logLevelString);
-
-            // timestamp
-            var timestampFormat = Options.TimestampFormat;
-            if (timestampFormat != null)
-            {
-                var dateTime = GetCurrentDateTime();
-                logBuilder.Append(dateTime.ToString(timestampFormat));
-            }
-
-            // category and event id
-            logBuilder.Append(logName);
-            logBuilder.Append('[');
-            logBuilder.Append(eventId);
-            logBuilder.Append(']');
-
-            // scope information
-            GetScopeInformation(logBuilder, multiLine: false);
-
-            // message
-            if (!string.IsNullOrEmpty(message))
-            {
-                logBuilder.Append(' ');
-                // message
-                AppendAndReplaceNewLine(logBuilder, message);
-            }
-
-            // exception
-            // System.InvalidOperationException at Namespace.Class.Function() in File:line X
-            if (exception != null)
-            {
-                logBuilder.Append(' ');
-                AppendAndReplaceNewLine(logBuilder, exception.ToString());
-            }
-
-            // newline delimiter
-            logBuilder.Append(Environment.NewLine);
-
-            return new LogMessageEntry(
-                message: logBuilder.ToString(),
-                logAsError: logLevel >= Options.LogToStandardErrorThreshold
-            );
-
-            static void AppendAndReplaceNewLine(StringBuilder sb, string message)
-            {
-                var len = sb.Length;
-                sb.Append(message);
-                sb.Replace(Environment.NewLine, " ", len, message.Length);
-            }
-        }
-
-        private DateTime GetCurrentDateTime()
-        {
-            return Options.UseUtcTimestamp ? DateTime.UtcNow : DateTime.Now;
-        }
 
         public bool IsEnabled(LogLevel logLevel)
         {
@@ -229,118 +65,5 @@ namespace Microsoft.Extensions.Logging.Console
         }
 
         public IDisposable BeginScope<TState>(TState state) => ScopeProvider?.Push(state) ?? NullScope.Instance;
-
-        private static string GetLogLevelString(LogLevel logLevel)
-        {
-            switch (logLevel)
-            {
-                case LogLevel.Trace:
-                    return "trce";
-                case LogLevel.Debug:
-                    return "dbug";
-                case LogLevel.Information:
-                    return "info";
-                case LogLevel.Warning:
-                    return "warn";
-                case LogLevel.Error:
-                    return "fail";
-                case LogLevel.Critical:
-                    return "crit";
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(logLevel));
-            }
-        }
-
-        private static string GetSyslogSeverityString(LogLevel logLevel)
-        {
-            // 'Syslog Message Severities' from https://tools.ietf.org/html/rfc5424.
-            switch (logLevel)
-            {
-                case LogLevel.Trace:
-                case LogLevel.Debug:
-                    return "<7>"; // debug-level messages
-                case LogLevel.Information:
-                    return "<6>"; // informational messages
-                case LogLevel.Warning:
-                    return "<4>"; // warning conditions
-                case LogLevel.Error:
-                    return "<3>"; // error conditions
-                case LogLevel.Critical:
-                    return "<2>"; // critical conditions
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(logLevel));
-            }
-        }
-
-        private ConsoleColors GetLogLevelConsoleColors(LogLevel logLevel)
-        {
-            if (Options.DisableColors)
-            {
-                return new ConsoleColors(null, null);
-            }
-
-            // We must explicitly set the background color if we are setting the foreground color,
-            // since just setting one can look bad on the users console.
-            switch (logLevel)
-            {
-                case LogLevel.Critical:
-                    return new ConsoleColors(ConsoleColor.White, ConsoleColor.Red);
-                case LogLevel.Error:
-                    return new ConsoleColors(ConsoleColor.Black, ConsoleColor.Red);
-                case LogLevel.Warning:
-                    return new ConsoleColors(ConsoleColor.Yellow, ConsoleColor.Black);
-                case LogLevel.Information:
-                    return new ConsoleColors(ConsoleColor.DarkGreen, ConsoleColor.Black);
-                case LogLevel.Debug:
-                    return new ConsoleColors(ConsoleColor.Gray, ConsoleColor.Black);
-                case LogLevel.Trace:
-                    return new ConsoleColors(ConsoleColor.Gray, ConsoleColor.Black);
-                default:
-                    return new ConsoleColors(DefaultConsoleColor, DefaultConsoleColor);
-            }
-        }
-
-        private void GetScopeInformation(StringBuilder stringBuilder, bool multiLine)
-        {
-            var scopeProvider = ScopeProvider;
-            if (Options.IncludeScopes && scopeProvider != null)
-            {
-                var initialLength = stringBuilder.Length;
-
-                scopeProvider.ForEachScope((scope, state) =>
-                {
-                    var (builder, paddAt) = state;
-                    var padd = paddAt == builder.Length;
-                    if (padd)
-                    {
-                        builder.Append(_messagePadding);
-                        builder.Append("=> ");
-                    }
-                    else
-                    {
-                        builder.Append(" => ");
-                    }
-                    builder.Append(scope);
-                }, (stringBuilder, multiLine ? initialLength : -1));
-
-                if (stringBuilder.Length > initialLength && multiLine)
-                {
-                    stringBuilder.AppendLine();
-                }
-            }
-        }
-
-        private readonly struct ConsoleColors
-        {
-            public ConsoleColors(ConsoleColor? foreground, ConsoleColor? background)
-            {
-                Foreground = foreground;
-                Background = background;
-            }
-
-            public ConsoleColor? Foreground { get; }
-
-            public ConsoleColor? Background { get; }
-        }
     }
 }
