+++ src/libraries/Microsoft.Extensions.Logging.Console/src/Formatter/DefaultLogFormatterOptions.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Extensions.Logging.Console
+{
+    public class DefaultLogFormatterOptions
+    {
+
+    }
+}
