+++ src/libraries/Microsoft.Extensions.Logging.Console/src/Formatter/SystemdLogFormatterOptions.cs
@@ -1,0 +1,9 @@
+using System;
+using System.Text;
+
+namespace Microsoft.Extensions.Logging.Console
+{
+    public class SystemdLogFormatterOptions
+    {
+    }
+}
