+++ src/libraries/Microsoft.Extensions.Logging/tests/Common/ConsoleLoggerTest.cs
@@ -21,6 +21,7 @@ namespace Microsoft.Extensions.Logging.Test
         private const string _loggerName = "test";
         private const string _state = "This is a test, and {curly braces} are just fine!";
         private readonly Func<object, Exception, string> _defaultFormatter = (state, exception) => state.ToString();
+        private readonly IEnumerable<ILogFormatter> _formatters = new ILogFormatter[] { new DefaultLogFormatter { Options = new ConsoleLoggerOptions() }, new SystemdLogFormatter { Options = new ConsoleLoggerOptions() } };
 
         private static (ConsoleLogger Logger, ConsoleSink Sink, ConsoleSink ErrorSink, Func<LogLevel, string> GetLevelPrefix, int WritesPerMsg) SetUp(ConsoleLoggerOptions options = null)
         {
@@ -33,9 +34,24 @@ namespace Microsoft.Extensions.Logging.Test
             consoleLoggerProcessor.Console = console;
             consoleLoggerProcessor.ErrorConsole = errorConsole;
 
-            var logger = new ConsoleLogger(_loggerName, consoleLoggerProcessor);
-            logger.ScopeProvider = new LoggerExternalScopeProvider();
-            logger.Options = options ?? new ConsoleLoggerOptions();
+            var testOptions = options ?? new ConsoleLoggerOptions();
+            var scopeProvider = new LoggerExternalScopeProvider();
+            var formatters = new ILogFormatter[]
+            {
+                new DefaultLogFormatter
+                {
+                    Options = testOptions,
+                    ScopeProvider = scopeProvider
+                },
+                new SystemdLogFormatter
+                {
+                    Options = testOptions,
+                    ScopeProvider = scopeProvider
+                }
+            };
+            var logger = new ConsoleLogger(_loggerName, consoleLoggerProcessor, formatters);
+            logger.ScopeProvider = scopeProvider;
+            logger.Options = testOptions;
             Func<LogLevel, string> levelAsString;
             int writesPerMsg;
             switch (logger.Options.Format)
@@ -994,7 +1010,7 @@ namespace Microsoft.Extensions.Logging.Test
             var processor = new ConsoleLoggerProcessor();
             processor.Console = console;
 
-            var logger = new ConsoleLogger(_loggerName, loggerProcessor: processor);
+            var logger = new ConsoleLogger(_loggerName, loggerProcessor: processor, _formatters);
             logger.Options = new ConsoleLoggerOptions();
             // Act
             processor.Dispose();
@@ -1023,7 +1039,7 @@ namespace Microsoft.Extensions.Logging.Test
         {
             // Arrange
             var monitor = new TestOptionsMonitor(new ConsoleLoggerOptions() { DisableColors = true });
-            var loggerProvider = new ConsoleLoggerProvider(monitor);
+            var loggerProvider = new ConsoleLoggerProvider(monitor, _formatters);
             var logger = (ConsoleLogger)loggerProvider.CreateLogger("Name");
 
             // Act & Assert
@@ -1054,7 +1070,7 @@ namespace Microsoft.Extensions.Logging.Test
         {
             // Arrange
             var monitor = new TestOptionsMonitor(new ConsoleLoggerOptions());
-            var loggerProvider = new ConsoleLoggerProvider(monitor);
+            var loggerProvider = new ConsoleLoggerProvider(monitor, _formatters);
             var logger = (ConsoleLogger)loggerProvider.CreateLogger("Name");
 
             // Act & Assert
@@ -1084,7 +1100,7 @@ namespace Microsoft.Extensions.Logging.Test
         public void ConsoleLoggerOptions_TimeStampFormat_MultipleReloads()
         {
             var monitor = new TestOptionsMonitor(new ConsoleLoggerOptions());
-            var loggerProvider = new ConsoleLoggerProvider(monitor);
+            var loggerProvider = new ConsoleLoggerProvider(monitor, _formatters);
             var logger = (ConsoleLogger)loggerProvider.CreateLogger("Name");
 
             Assert.Null(logger.Options.TimestampFormat);
@@ -1099,7 +1115,7 @@ namespace Microsoft.Extensions.Logging.Test
         {
             // Arrange
             var monitor = new TestOptionsMonitor(new ConsoleLoggerOptions() { IncludeScopes = true });
-            var loggerProvider = new ConsoleLoggerProvider(monitor);
+            var loggerProvider = new ConsoleLoggerProvider(monitor, _formatters);
             var logger = (ConsoleLogger)loggerProvider.CreateLogger("Name");
 
             // Act & Assert
@@ -1130,7 +1146,7 @@ namespace Microsoft.Extensions.Logging.Test
         {
             // Arrange
             var monitor = new TestOptionsMonitor(new ConsoleLoggerOptions());
-            var loggerProvider = new ConsoleLoggerProvider(monitor);
+            var loggerProvider = new ConsoleLoggerProvider(monitor, _formatters);
             var logger = (ConsoleLogger)loggerProvider.CreateLogger("Name");
 
             // Act & Assert
@@ -1144,7 +1160,7 @@ namespace Microsoft.Extensions.Logging.Test
         {
             // Arrange
             var monitor = new TestOptionsMonitor(new ConsoleLoggerOptions());
-            var loggerProvider = new ConsoleLoggerProvider(monitor);
+            var loggerProvider = new ConsoleLoggerProvider(monitor, _formatters);
             var logger = (ConsoleLogger)loggerProvider.CreateLogger("Name");
 
             // Act & Assert
